refactor: _components를 (main) 스코프로 이동
- app/_components/를 app/(main)/_components/로 이동
- 모든 사용처가 (main)/ 하위이므로 스코프에 맞게 위치 조정
- 관련 import 경로 2곳 수정
