+++ src/app/(main)/_components/home-footer.tsx
