+++ src/app/(main)/_components/home-hero-section.tsx
