+++ src/app/(main)/_components/home-promo-section.tsx
